% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_robust.R
\name{augmented_model_matrix.lm_robust}
\alias{augmented_model_matrix.lm_robust}
\title{Same as model.matrix.lm
***removed export tag}
\usage{
\method{augmented_model_matrix}{lm_robust}(obj, cluster, inverse_var, ignore_FE)
}
\description{
Same as model.matrix.lm
***removed export tag
}
