% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{MortalityRates}
\alias{MortalityRates}
\title{State-level annual mortality rates by cause among 18-20 year-olds}
\format{
A data frame with 5508 rows and 12 variables: \describe{ 
  \item{year}{Year of observation} 
  \item{state}{identifier for state} 
  \item{count}{Number of deaths} 
  \item{pop}{Population size} 
  \item{legal}{Proportion of 18-20 year-old population that is legally allowed to drink} 
  \item{beertaxa}{Beer taxation rate} 
  \item{beerpercap}{Beer consumption per capita} 
  \item{winepercap}{Wine consumption per capita} 
  \item{spiritpercap}{Spirits consumption per capita} 
  \item{totpercap}{Total alcohol consumption per capita} 
  \item{mrate}{Mortality rate per 10,000} 
  \item{cause}{Cause of death} 
  }
}
\source{
\href{http://masteringmetrics.com/wp-content/uploads/2015/01/deaths.dta}{Mastering
  'Metrics data archive}
}
\usage{
MortalityRates
}
\description{
A dataset containing state-level annual mortality rates for select causes of
death, as well as data related to the minimum legal drinking age and alcohol
consumption.
}
\references{
Angrist, J. D., and Pischke, J. S. (2014). _Mastering'metrics: the path from
cause to effect_. Princeton University Press, 2014.

Carpenter, C., & Dobkin, C. (2011). The minimum legal drinking age and public
health. _Journal of Economic Perspectives, 25_(2), 133-156.
\doi{10.1257/jep.25.2.133}
}
\keyword{datasets}
