% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{v_scale.mmrm}
\alias{v_scale.mmrm}
\title{Scaling constant for mmrm objects}
\usage{
\method{v_scale}{mmrm}(obj)
}
\arguments{
\item{obj}{A fitted \code{\link[mmrm]{mmrm}} model object.}
}
\value{
A scalar equal to the number of subjects.
}
\description{
Returns the scaling constant for the sandwich estimator,
  which for \code{\link[mmrm]{mmrm}} models is the number of subjects
  (clusters).
}
