% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest.R
\name{vcovCR.fixest}
\alias{vcovCR.fixest}
\title{Cluster-robust variance-covariance matrix for a fixest object.}
\usage{
\method{vcovCR}{fixest}(
  obj,
  cluster,
  type,
  target = NULL,
  inverse_var = NULL,
  form = "sandwich",
  ...
)
}
\arguments{
\item{obj}{Fitted model for which to calculate the variance-covariance matrix}

\item{cluster}{Optional expression or vector indicating which observations
belong to the same cluster. If not specified, will be set to the first
fixed effect variable from the model formula if available.}

\item{type}{Character string specifying which small-sample adjustment should 
be used, with available options \code{"CR0"}, \code{"CR1"}, \code{"CR1p"},
\code{"CR1S"}, \code{"CR2"}, or \code{"CR3"}. See "Details" section of 
\code{\link{vcovCR}} for further information.}

\item{target}{Optional matrix or vector describing the working
variance-covariance model used to calculate the \code{CR2} and \code{CR4}
adjustment matrices. If a vector, the target matrix is assumed to be
diagonal. If not specified, the target is taken to be an identity matrix.}

\item{inverse_var}{Optional logical indicating whether the weights used in 
fitting the model are inverse-variance. If not specified, \code{vcovCR} 
will attempt to infer a value.}

\item{form}{Controls the form of the returned matrix. The default 
\code{"sandwich"} will return the sandwich variance-covariance matrix. 
Alternately, setting \code{form = "meat"} will return only the meat of the 
sandwich and setting \code{form = B}, where \code{B} is a matrix of 
appropriate dimension, will return the sandwich variance-covariance matrix 
calculated using \code{B} as the bread. \code{form = "estfun"} will return the 
(appropriately scaled) estimating function, the transposed crossproduct of 
which is equal to the sandwich variance-covariance matrix.}

\item{...}{Additional arguments available for some classes of objects.}
}
\value{
An object of class \code{c("vcovCR","clubSandwich")}, which consists
  of a matrix of the estimated variance of and covariances between the
  regression coefficient estimates.
}
\description{
\code{vcovCR} returns a sandwich estimate of the variance-covariance matrix
of a set of regression coefficient estimates from a
\code{\link[fixest]{feols}} object.
}
\examples{

if (requireNamespace("fixest", quietly = TRUE)) withAutoprint({

  library(fixest)
  data(trade, package = "fixest")

  # feols model with Origin and Destination fixed effects
  fit <- feols(log(Euros) ~ log(dist_km) | Origin + Destination,
               data = trade)

  # CR2 cluster-robust variance estimator (cluster auto-detected from FE)
  vcovCR(fit, type = "CR2")

  # Coefficient tests with Satterthwaite degrees of freedom
  coef_test(fit, vcov = "CR2", test = "Satterthwaite")

  # Explicitly specify clustering variable
  vcovCR(fit, cluster = trade$Origin, type = "CR2")

})

}
\seealso{
\code{\link{vcovCR}}
}
