% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{vcovCR.mmrm}
\alias{vcovCR.mmrm}
\title{Cluster-robust variance-covariance matrix for an mmrm object.}
\usage{
\method{vcovCR}{mmrm}(obj, cluster, type, target, inverse_var, form = "sandwich", ...)
}
\arguments{
\item{obj}{Fitted model for which to calculate the variance-covariance matrix}

\item{cluster}{Optional expression or vector indicating which observations
belong to the same cluster. If not specified, will be set to the subject
variable from the \code{mmrm} model formula.}

\item{type}{Character string specifying which small-sample adjustment should 
be used, with available options \code{"CR0"}, \code{"CR1"}, \code{"CR1p"},
\code{"CR1S"}, \code{"CR2"}, or \code{"CR3"}. See "Details" section of 
\code{\link{vcovCR}} for further information.}

\item{target}{Optional matrix or vector describing the working
variance-covariance model used to calculate the \code{CR2} and \code{CR4}
adjustment matrices. If not specified, the target is taken to be the
estimated variance-covariance structure of the \code{mmrm} object, with
weights applied if present.}

\item{inverse_var}{Optional logical indicating whether the weights used in 
fitting the model are inverse-variance. If not specified, \code{vcovCR} 
will attempt to infer a value.}

\item{form}{Controls the form of the returned matrix. The default 
\code{"sandwich"} will return the sandwich variance-covariance matrix. 
Alternately, setting \code{form = "meat"} will return only the meat of the 
sandwich and setting \code{form = B}, where \code{B} is a matrix of 
appropriate dimension, will return the sandwich variance-covariance matrix 
calculated using \code{B} as the bread. \code{form = "estfun"} will return the 
(appropriately scaled) estimating function, the transposed crossproduct of 
which is equal to the sandwich variance-covariance matrix.}

\item{...}{Additional arguments available for some classes of objects.}
}
\value{
An object of class \code{c("vcovCR","clubSandwich")}, which consists
  of a matrix of the estimated variance of and covariances between the
  regression coefficient estimates.
}
\description{
\code{vcovCR} returns a sandwich estimate of the variance-covariance matrix
of a set of regression coefficient estimates from an
\code{\link[mmrm]{mmrm}} object.
}
\examples{

if (requireNamespace("mmrm", quietly = TRUE)) withAutoprint({

  library(mmrm)
  data(fev_data, package = "mmrm")

  # Fit an mmrm model with unstructured covariance
  mmrm_fit <- mmrm(
    FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
    data = fev_data
  )

  # CR2 cluster-robust variance estimator (cluster auto-detected)
  vcovCR(mmrm_fit, type = "CR2")

  # Coefficient tests with Satterthwaite degrees of freedom
  coef_test(mmrm_fit, vcov = "CR2", test = "Satterthwaite")

  # Fit a weighted mmrm model
  fev_data$wt <- 1 + rpois(nrow(fev_data), lambda = 3)
  mmrm_wt <- mmrm(
    FEV1 ~ RACE + SEX + ARMCD + us(AVISIT | USUBJID),
    data = fev_data,
    weights = fev_data$wt
  )

  # CR2 works with weighted models
  vcovCR(mmrm_wt, type = "CR2")

})

}
\seealso{
\code{\link{vcovCR}}
}
