% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plm.R
\name{vcovCR.plm}
\alias{vcovCR.plm}
\title{Cluster-robust variance-covariance matrix for a plm object.}
\usage{
\method{vcovCR}{plm}(obj, cluster, type, target, inverse_var)
}
\arguments{
\item{obj}{Fitted model for which to calcualte the variance-covariance matrix}

\item{cluster}{Optional character string, expression, or vector indicating
which observations belong to the same cluster. For fixed-effect models that
include individual effects or time effects (but not both), the cluster will
be taken equal to the included fixed effects if not otherwise specified.
Clustering on individuals can also be obtained by taking \code{cluster =
"individual"} and clustering on time periods can be obtained with
\code{cluster = "time"}. For random-effects models, the cluster will be
taken equal to the included random effect identifier if not otherwise
specified.}

\item{type}{Character string specifying which small-sample adjustment should
be used.}

\item{target}{Optional matrix or vector describing the working
variance-covariance model used to calculate the \code{CR2} and \code{CR4}
adjustment matrices. By default, the target is taken to be an identity
matrix for fixed effect models or the estimated compound-symmetric covariance matrix for random effects models.}

\item{inverse_var}{Optional logical indicating whether the weights used in
  fitting the model are inverse-variance. If not specified, \code{vcovCR}
  will attempt to infer a value.}
}
\value{
An object of class \code{c("vcovCR","clubSandwich")}, which consists
  of a matrix of the estimated variance of and covariances between the
  regression coefficient estimates.
}
\description{
\code{vcovCR} returns a sandwich estimate of the variance-covariance matrix
of a set of regression coefficient estimates from a \code{\link[plm]{plm}}
object.
}
\seealso{
\code{\link{vcovCR}}
}

