% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_robust.R
\name{vcovCR.iv_robust}
\alias{vcovCR.iv_robust}
\title{Cluster-robust variance-covariance matrix for an
\code{estimatr::iv_robust} object.}
\usage{
\method{vcovCR}{iv_robust}(
  obj,
  cluster,
  type,
  target = NULL,
  inverse_var = FALSE,
  form = "sandwich",
  ...
)
}
\arguments{
\item{obj}{Fitted model for which to calculate the variance-covariance matrix}

\item{cluster}{Expression or vector indicating which observations belong to
the same cluster. Required for \code{iv_robust} objects unless the model
was fitted with a \code{clusters} argument.}

\item{type}{Character string specifying which small-sample adjustment should 
be used, with available options \code{"CR0"}, \code{"CR1"}, \code{"CR1p"},
\code{"CR1S"}, \code{"CR2"}, or \code{"CR3"}. See "Details" section of 
\code{\link{vcovCR}} for further information.}

\item{target}{Optional matrix or vector describing the working
variance-covariance model used to calculate the \code{CR2} and \code{CR4}
adjustment matrices. If a vector, the target matrix is assumed to be
diagonal. If not specified, the target is taken to be an identity matrix.}

\item{inverse_var}{Not used for \code{iv_robust} objects.}

\item{form}{Controls the form of the returned matrix. The default 
\code{"sandwich"} will return the sandwich variance-covariance matrix. 
Alternately, setting \code{form = "meat"} will return only the meat of the 
sandwich and setting \code{form = B}, where \code{B} is a matrix of 
appropriate dimension, will return the sandwich variance-covariance matrix 
calculated using \code{B} as the bread. \code{form = "estfun"} will return the 
(appropriately scaled) estimating function, the transposed crossproduct of 
which is equal to the sandwich variance-covariance matrix.}

\item{...}{Additional arguments available for some classes of objects.}
}
\value{
An object of class \code{c("vcovCR","clubSandwich")}, which consists
  of a matrix of the estimated variance of and covariances between the
  regression coefficient estimates.
}
\description{
\code{vcovCR} returns a sandwich estimate of the variance-covariance matrix
of a set of regression coefficient estimates from an
\code{\link[estimatr]{iv_robust}} object.
}
\examples{

if (requireNamespace("estimatr", quietly = TRUE)) withAutoprint({
  library(estimatr)

  set.seed(20261201)
  N <- 200
  cluster <- factor(rep(1:20, each = 10))
  z1 <- rnorm(N); z2 <- rnorm(N); x_exog <- rnorm(N)
  x_endog <- 0.5 * z1 + 0.3 * z2 + rnorm(N)
  y <- 1 + 2 * x_endog + 0.5 * x_exog + rnorm(N)
  dat <- data.frame(y, x_endog, x_exog, z1, z2, cluster)

  # Basic 2SLS with clusters and CR2
  iv_fit <- iv_robust(
    y ~ x_endog + x_exog | x_exog + z1 + z2,
    data = dat, clusters = cluster, se_type = "CR2"
  )
  vcovCR(iv_fit)
  conf_int(iv_fit, vcov = "CR2")

  # 2SLS with fixed effects
  iv_fe <- iv_robust(
    y ~ x_endog + x_exog | x_exog + z1 + z2,
    data = dat, clusters = cluster,
    fixed_effects = ~ cluster, se_type = "CR2"
  )
  vcovCR(iv_fe)

})

}
\seealso{
\code{\link{vcovCR}}
}
