% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-uni.R
\name{vcovCR.rma.uni}
\alias{vcovCR.rma.uni}
\title{Cluster-robust variance-covariance matrix for a rma.uni object.}
\usage{
\method{vcovCR}{rma.uni}(obj, cluster, type, target, inverse_var,
  form = "sandwich", ...)
}
\arguments{
\item{obj}{Fitted model for which to calcualte the variance-covariance matrix}

\item{cluster}{Expression or vector indicating which observations 
belong to the same cluster. Required for \code{rma.uni} objects.}

\item{type}{Character string specifying which small-sample adjustment should 
be used.}

\item{target}{Optional matrix or vector describing the working 
variance-covariance model used to calculate the \code{CR2} and \code{CR4} 
adjustment matrices. If not specified, the target is taken to be diagonal
with entries equal to the estimated marginal variance of the effect sizes.}

\item{inverse_var}{Optional logical indicating whether the weights used in 
fitting the model are inverse-variance. If not specified, \code{vcovCR} 
will attempt to infer a value.}

\item{form}{Controls the form of the returned matrix. The default 
\code{"sandwich"} will return the sandwich variance-covariance matrix. 
Alternately, setting \code{form = "meat"} will return only the meat of the 
sandwich and setting \code{form = B}, where \code{B} is a matrix of 
appropriate dimension, will return the sandwich variance-covariance matrix 
calculated using \code{B} as the bread.}

\item{...}{Additional arguments available for some classes of objects.}
}
\value{
An object of class \code{c("vcovCR","clubSandwich")}, which consists 
  of a matrix of the estimated variance of and covariances between the 
  regression coefficient estimates.
}
\description{
\code{vcovCR} returns a sandwich estimate of the variance-covariance matrix 
of a set of regression coefficient estimates from a 
\code{\link[metafor]{rma.uni}} object.
}
\examples{
library(metafor)
data(corrdat, package = "robumeta")

mfor_fit <- rma.uni(effectsize ~ males + college + binge,
                     vi = var, data = corrdat, method = "FE")
mfor_fit
mfor_CR2 <- vcovCR(mfor_fit, type = "CR2", cluster = corrdat$studyid)
mfor_CR2
coef_test(mfor_fit, vcov = mfor_CR2, test = c("Satterthwaite", "saddlepoint"))
Wald_test(mfor_fit, constraints = 2:4, vcov = mfor_CR2)

}
\seealso{
\code{\link{vcovCR}}
}

